+++ luasrc/model/cbi/serverchan/log.lua
@@ -1,0 +1,5 @@
+f = SimpleForm("serverchan")
+f.reset = false
+f.submit = false
+f:append(Template("serverchan/log"))
+return f
