+++ root/etc/uci-defaults/luci-serverchan
@@ -1,0 +1,12 @@
+#!/bin/sh
+touch /etc/config/serverchan
+
+uci -q batch <<-EOF >/dev/null
+	delete ucitrack.@serverchan[-1]
+	add ucitrack serverchan
+	set ucitrack.@serverchan[-1].exec='/etc/init.d/serverchan restart'
+	commit ucitrack
+EOF
+# remove LuCI cache
+rm -f /tmp/luci-indexcache
+exit 0
